added event listener to groundoverlaymanager
