+++ maps-compose-utils/src/main/java/com/google/maps/android/compose/kml/event/KmlEvent.kt
@@ -1,0 +1,19 @@
+package com.google.maps.android.compose.kml.event
+
+import com.google.maps.android.compose.kml.manager.GroundOverlayManager
+import com.google.maps.android.compose.kml.manager.MarkerManager
+import com.google.maps.android.compose.kml.manager.PolylineManager
+
+public sealed class KmlEvent {
+    public sealed class Marker : KmlEvent() {
+        public data class Clicked(val data: MarkerManager.MarkerProperties) : Marker()
+    }
+
+    public sealed class Polyline : KmlEvent() {
+        public data class Clicked(val data: PolylineManager.PolylineProperties) : Polyline()
+    }
+
+    public sealed class GroundOverlay : KmlEvent() {
+        public data class Clicked(val data: GroundOverlayManager.GroundOverlayProperties) : GroundOverlay()
+    }
+}
